+++ src/main/kotlin/blackjack/domain/deck/Deck.kt
@@ -1,0 +1,9 @@
+package blackjack.domain.deck
+
+import blackjack.domain.TrumpCard
+
+interface Deck {
+    fun pop(): TrumpCard
+
+    fun makeCards(): List<TrumpCard>
+}
