+++ src/main/kotlin/blackjack/domain/deck/ShuffledDeck.kt
@@ -1,22 +1,21 @@
-package blackjack.domain
+package blackjack.domain.deck
 
+import blackjack.domain.CardTier
+import blackjack.domain.Shape
+import blackjack.domain.TrumpCard
 import java.util.ArrayDeque
 import java.util.Deque
 
-class Deck {
+class ShuffledDeck : Deck {
     private val cards: Deque<TrumpCard> = ArrayDeque()
 
     init {
-        getAllCards()
-    }
-
-    fun pop(): TrumpCard = cards.pop()
-
-    private fun getAllCards() {
         cards.addAll(makeCards())
     }
 
-    private fun makeCards(): List<TrumpCard> =
+    override fun pop(): TrumpCard = cards.pop()
+
+    override fun makeCards(): List<TrumpCard> =
         Shape.entries
             .flatMap { shape ->
                 CardTier.entries.map { tier ->
