+++ src/test/kotlin/blackjack/BlackJackGameTest.kt
@@ -2,10 +2,10 @@ package blackjack
 
 import blackjack.domain.BlackJackGame
 import blackjack.domain.CardTier
-import blackjack.domain.Deck
 import blackjack.domain.Shape
 import blackjack.domain.TrumpCard
 import blackjack.domain.UserChoice
+import blackjack.domain.deck.ShuffledDeck
 import blackjack.fixture.playersFixture
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -14,7 +14,7 @@ class BlackJackGameTest {
     @Test
     fun `게임을 시작하면 각 플레이어와 딜러는 2장의 카드를 지급받는다`() {
         val players = playersFixture()
-        val game = BlackJackGame(players, Deck())
+        val game = BlackJackGame(players, ShuffledDeck())
         game.handOutInitializedCards(2)
         assertThat(players.first().cards.size).isEqualTo(2)
     }
@@ -22,7 +22,7 @@ class BlackJackGameTest {
     @Test
     fun `플레이어가 hit을 선택하면 카드를 한 장 추가한다`() {
         val players = playersFixture()
-        val game = BlackJackGame(players, Deck())
+        val game = BlackJackGame(players, ShuffledDeck())
         players.first().addCard(TrumpCard(CardTier.JACK, Shape.DIA))
         players.first().addCard(TrumpCard(CardTier.JACK, Shape.DIA))
         game.playGame(
@@ -35,7 +35,7 @@ class BlackJackGameTest {
     @Test
     fun `플레이어가 stay를 선택하면 카드의 장수가 유지된다`() {
         val players = playersFixture()
-        val game = BlackJackGame(players, Deck())
+        val game = BlackJackGame(players, ShuffledDeck())
 
         game.playGame(
             getPlayerChoice = { UserChoice.from("n") },
